add test cases and edit jacoco plugin
