+++ component/src/main/java/org/wso2/extension/siddhi/io/email/source/EmailMessageProcessor.java
@@ -91,18 +91,11 @@ public class EmailMessageProcessor implements CarbonMessageProcessor {
         return values;
     }
 
+    @Override public void setTransportSender(TransportSender transportSender) {}
 
-    @Override public void setTransportSender(TransportSender transportSender) {
-
-    }
-
-    @Override public void setClientConnector(ClientConnector clientConnector) {
-
-    }
-
+    @Override public void setClientConnector(ClientConnector clientConnector) {}
 
     @Override public String getId() {
         return "email-message-processor";
     }
-
 }
