+++ component/src/main/java/org/wso2/extension/siddhi/io/email/source/EmailSource.java
@@ -676,7 +676,6 @@ public class EmailSource extends Source {
         String sslEnable = optionHolder.validateAndGetStaticValue(EmailConstants.EMAIL_RECEIVER_SSL_ENABLE, configReader
                 .readConfig(EmailConstants.EMAIL_RECEIVER_SSL_ENABLE,
                         EmailConstants.EMAIL_RECEIVER_DEFAULT_SSL_ENABLE));
-
         if (!(sslEnable.equalsIgnoreCase("true") || sslEnable.equalsIgnoreCase("false"))) {
             throw new SiddhiAppCreationException(
                     EmailConstants.STORE + "could be either 'true' " + "or 'false'. But found: " + sslEnable);
@@ -685,16 +684,15 @@ public class EmailSource extends Source {
 
         String port = optionHolder.validateAndGetStaticValue(EmailConstants.EMAIL_RECEIVER_PORT, configReader
                 .readConfig(EmailConstants.EMAIL_RECEIVER_PORT, EmailConstants.EMPTY_STRING));
-
         if (port.isEmpty()) {
-        if (sslEnable.equalsIgnoreCase("true") && store.equalsIgnoreCase("imap")) {
-                port = EmailConstants.EMAIL_RECEIVER_DEFAULT_PORT;
-        } else {
-                throw new SiddhiAppCreationException("Default value for the port can be only used if 'ssl.enable'"
-                        + " is 'true' and store type is 'imap' only.");
+                if (sslEnable.equalsIgnoreCase("true") && store.equalsIgnoreCase("imap")) {
+                        port = EmailConstants.EMAIL_RECEIVER_DEFAULT_PORT;
+                } else {
+                         throw new SiddhiAppCreationException(
+                                    "Default value for the port can be only used if 'ssl.enable'"
+                                            + " is 'true' and store type is 'imap' only.");
+                }
         }
-        }
-
         properties.put("mail." + store + ".port", port);
 
         String pollingInterval = optionHolder.validateAndGetStaticValue(EmailConstants.POLLING_INTERVAL,
@@ -705,9 +703,7 @@ public class EmailSource extends Source {
         //get a list of valid search term keys.
         List<String> validSearchTermKeys = Stream.of(EmailConstants.SearchTermKeys.values()).
                 map(EmailConstants.SearchTermKeys::name).collect(Collectors.toList());
-
         List<String> givenSearchtermkeys = new ArrayList<>();
-
         String searchTerm = optionHolder.validateAndGetStaticValue(EmailConstants.EMAIL_SEARCH_TERM,
                 configReader.readConfig(EmailConstants.EMAIL_SEARCH_TERM, EmailConstants.EMPTY_STRING));
 
@@ -730,7 +726,6 @@ public class EmailSource extends Source {
                     }
                 }
             }
-
             //check given search term keys are valid.
             if (!validSearchTermKeys.containsAll(givenSearchtermkeys)) {
                 throw new SiddhiAppCreationException("Valid search term to search emails are" +
