+++ component/src/test/resources/testng.xml
@@ -24,6 +24,8 @@
         <classes>
             <class name="org.wso2.extension.siddhi.io.email.sink.EmailSinkTestCase"/>
             <class name="org.wso2.extension.siddhi.io.email.source.EmailSourceImapTestCase"/>
+            <class name="org.wso2.extension.siddhi.io.email.source.EmailSourceActionAfterProcessedTestCase"/>
+            <class name="org.wso2.extension.siddhi.io.email.source.mailServerSslConnectionTestCase"/>
         </classes>
     </test>
 </suite>
