+++ pom.xml
@@ -111,6 +111,12 @@
                 <version>${com.icegreen.version}</version>
                 <scope>test</scope>
             </dependency>
+            <dependency>
+                <groupId>org.jacoco</groupId>
+                <artifactId>org.jacoco.agent</artifactId>
+                <classifier>runtime</classifier>
+                <version>${jacoco.version}</version>
+            </dependency>
         </dependencies>
 
     </dependencyManagement>
@@ -138,65 +144,19 @@
                 </executions>
             </plugin>
         </plugins>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-surefire-plugin</artifactId>
-                    <version>2.19.1</version>
-                    <configuration>
-                        <includes>
-                            <include>**/*TestCase.java</include>
-                        </includes>
-                        <reportFormat>brief</reportFormat>
-                        <useFile>false</useFile>
-                        <forkMode>once</forkMode>
-                        <argLine>${surefireArgLine} -ea -Xmx512m</argLine>
-                    </configuration>
-                </plugin>
-                <plugin>
-                    <groupId>org.jacoco</groupId>
-                    <artifactId>jacoco-maven-plugin</artifactId>
-                    <version>${jacoco.maven.version}</version>
-                    <executions>
-                        <execution>
-                            <id>jacoco-initialize</id>
-                            <goals>
-                                <goal>prepare-agent</goal>
-                            </goals>
-                            <configuration>
-                                <destFile>${basedir}/target/coverage-reports/jacoco.exec</destFile>
-                                <propertyName>surefireArgLine</propertyName>
-                            </configuration>
-                        </execution>
-                        <execution>
-                            <id>jacoco-site</id>
-                            <phase>post-integration-test</phase>
-                            <goals>
-                                <goal>report</goal>
-                            </goals>
-                            <configuration>
-                                <dataFile>${basedir}/target/coverage-reports/jacoco.exec</dataFile>
-                                <outputDirectory>${basedir}/target/coverage-reports/</outputDirectory>
-                            </configuration>
-                        </execution>
-                    </executions>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
+  </build>
     <properties>
         <siddhi.version>4.0.0-M52</siddhi.version>
         <log4j.version>1.2.17.wso2v1</log4j.version>
         <testng.version>6.8</testng.version>
         <!--<carbon.feature.plugin.version>3.0.0</carbon.feature.plugin.version>-->
-        <jacoco.maven.version>0.7.8</jacoco.maven.version>
+        <jacoco.version>0.7.9</jacoco.version>
         <carbon.transport.version>5.0.0</carbon.transport.version>
         <com.icegreen.version>1.5.5</com.icegreen.version>
         <text.mapper.version>4.0.0-M7</text.mapper.version>
         <com.sun.mail.version>1.5.6</com.sun.mail.version>
         <siddhi.map.json.version>4.0.3</siddhi.map.json.version>
-        <xml.mapper.version>4.0.0-M8</xml.mapper.version>
+        <xml.mapper.version>4.0.0-M10</xml.mapper.version>
         <carbon.messaging.version>3.0.0</carbon.messaging.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
